Retire lib/data.ts: colocate report data with its consumers (Sprint 4)
Audit surprise: there was no actual dead code in lib/data.ts. All 10
remaining exports were imported by exactly one of two report pages:

  app/reports/feb-2026/page.tsx                  ← projects, projectTotals,
                                                    keyMetrics,
                                                    repositoryTimeline,
                                                    methodologyNote
  app/reports/presidential-brief-feb-2026/...    ← oredProjects,
                                                    shadowApplications,
                                                    currentConstraints,
                                                    recommendations,
                                                    adoptionPhases

The "Cleanup: drop 13 orphan exports" sweep (commit de18879) already
removed the truly-dead content. What remained was a misnamed file —
content that's just per-page data, not "legacy strategic data slated
for cleanup."

This PR colocates each set with its consumer:

  app/reports/feb-2026/data.ts                   (new)
  app/reports/presidential-brief-feb-2026/data.ts (new)
  lib/data.ts                                    (deleted)

Each report directory is now self-contained. The "slated for cleanup"
markers across CLAUDE.md, README.md, AGENTIC_CODING_HANDOUT.md,
CONTRIBUTING.md, and app/docs/architecture/page.tsx all retire.

Bonus sweep — found ~360 lines of dead .aispeg-deck CSS in
app/globals.css, theming the RevealDeck.tsx component dropped in
PR #90. Nothing referenced these styles anymore. Removed.

Also caught two PR #90 stragglers in CONTRIBUTING.md: the artifacts
description still listed "decks, talks", and the use-client list still
named RevealDeck.tsx. Both fixed.

Net: -397 lines, 10 files touched.
